+++ cmd/noms/noms_sync_test.go
@@ -27,7 +27,7 @@ type nomsSyncTestSuite struct {
 func (s *nomsSyncTestSuite) TestSyncValidation() {
 	sourceDB := datas.NewDatabase(nbs.NewLocalStore(s.DBDir, clienttest.DefaultMemTableSize))
 	source1 := sourceDB.GetDataset("src")
-	source1, err := sourceDB.CommitValue(source1, types.Number(42))
+	source1, err := sourceDB.CommitValue(source1, types.Float(42))
 	s.NoError(err)
 	source1HeadRef := source1.Head().Hash()
 	source1.Database().Close()
@@ -48,10 +48,10 @@ func (s *nomsSyncTestSuite) TestSync() {
 
 	sourceDB := datas.NewDatabase(nbs.NewLocalStore(s.DBDir, clienttest.DefaultMemTableSize))
 	source1 := sourceDB.GetDataset("src")
-	source1, err := sourceDB.CommitValue(source1, types.Number(42))
+	source1, err := sourceDB.CommitValue(source1, types.Float(42))
 	s.NoError(err)
 	source1HeadRef := source1.Head().Hash() // Remember first head, so we can sync to it.
-	source1, err = sourceDB.CommitValue(source1, types.Number(43))
+	source1, err = sourceDB.CommitValue(source1, types.Float(43))
 	s.NoError(err)
 	sourceDB.Close()
 
@@ -63,7 +63,7 @@ func (s *nomsSyncTestSuite) TestSync() {
 
 	db := datas.NewDatabase(nbs.NewLocalStore(s.DBDir2, clienttest.DefaultMemTableSize))
 	dest := db.GetDataset("dest")
-	s.True(types.Number(42).Equals(dest.HeadValue()))
+	s.True(types.Float(42).Equals(dest.HeadValue()))
 	db.Close()
 
 	// Pull from a dataset in one DB to an existing dataset in another
@@ -73,7 +73,7 @@ func (s *nomsSyncTestSuite) TestSync() {
 
 	db = datas.NewDatabase(nbs.NewLocalStore(s.DBDir2, clienttest.DefaultMemTableSize))
 	dest = db.GetDataset("dest")
-	s.True(types.Number(43).Equals(dest.HeadValue()))
+	s.True(types.Float(43).Equals(dest.HeadValue()))
 	db.Close()
 
 	// Pull when sink dataset is already up to date
@@ -87,7 +87,7 @@ func (s *nomsSyncTestSuite) TestSync() {
 
 	db = datas.NewDatabase(nbs.NewLocalStore(s.DBDir2, clienttest.DefaultMemTableSize))
 	dest = db.GetDataset("dest2")
-	s.True(types.Number(43).Equals(dest.HeadValue()))
+	s.True(types.Float(43).Equals(dest.HeadValue()))
 	db.Close()
 }
 
@@ -97,14 +97,14 @@ func (s *nomsSyncTestSuite) TestSync_Issue2598() {
 	sourceDB := datas.NewDatabase(nbs.NewLocalStore(s.DBDir, clienttest.DefaultMemTableSize))
 	// Create dataset "src1", which has a lineage of two commits.
 	source1 := sourceDB.GetDataset("src1")
-	source1, err := sourceDB.CommitValue(source1, types.Number(42))
+	source1, err := sourceDB.CommitValue(source1, types.Float(42))
 	s.NoError(err)
-	source1, err = sourceDB.CommitValue(source1, types.Number(43))
+	source1, err = sourceDB.CommitValue(source1, types.Float(43))
 	s.NoError(err)
 
 	// Create dataset "src2", with a lineage of one commit.
 	source2 := sourceDB.GetDataset("src2")
-	source2, err = sourceDB.CommitValue(source2, types.Number(1))
+	source2, err = sourceDB.CommitValue(source2, types.Float(1))
 	s.NoError(err)
 
 	sourceDB.Close() // Close Database backing both Datasets
@@ -116,7 +116,7 @@ func (s *nomsSyncTestSuite) TestSync_Issue2598() {
 
 	db := datas.NewDatabase(nbs.NewLocalStore(s.DBDir2, clienttest.DefaultMemTableSize))
 	dest := db.GetDataset("dest")
-	s.True(types.Number(43).Equals(dest.HeadValue()))
+	s.True(types.Float(43).Equals(dest.HeadValue()))
 	db.Close()
 
 	// Now, try syncing a second dataset. This crashed in issue #2598
@@ -126,7 +126,7 @@ func (s *nomsSyncTestSuite) TestSync_Issue2598() {
 
 	db = datas.NewDatabase(nbs.NewLocalStore(s.DBDir2, clienttest.DefaultMemTableSize))
 	dest = db.GetDataset("dest2")
-	s.True(types.Number(1).Equals(dest.HeadValue()))
+	s.True(types.Float(1).Equals(dest.HeadValue()))
 	db.Close()
 
 	sout, _ = s.MustRun(main, []string{"sync", sourceDataset, sinkDatasetSpec})
@@ -137,10 +137,10 @@ func (s *nomsSyncTestSuite) TestRewind() {
 	var err error
 	sourceDB := datas.NewDatabase(nbs.NewLocalStore(s.DBDir, clienttest.DefaultMemTableSize))
 	src := sourceDB.GetDataset("foo")
-	src, err = sourceDB.CommitValue(src, types.Number(42))
+	src, err = sourceDB.CommitValue(src, types.Float(42))
 	s.NoError(err)
 	rewindRef := src.HeadRef().TargetHash()
-	src, err = sourceDB.CommitValue(src, types.Number(43))
+	src, err = sourceDB.CommitValue(src, types.Float(43))
 	s.NoError(err)
 	sourceDB.Close() // Close Database backing both Datasets
 
@@ -150,6 +150,6 @@ func (s *nomsSyncTestSuite) TestRewind() {
 
 	db := datas.NewDatabase(nbs.NewLocalStore(s.DBDir, clienttest.DefaultMemTableSize))
 	dest := db.GetDataset("foo")
-	s.True(types.Number(42).Equals(dest.HeadValue()))
+	s.True(types.Float(42).Equals(dest.HeadValue()))
 	db.Close()
 }
